requirements.txt
numpy
tensorflow-gpu==1.7.0